Move tests into separate folder
