+++ src/core/mod.rs
@@ -2,3 +2,5 @@ pub mod components;
 mod contentline;
 pub mod parameters;
 pub mod properties;
+#[cfg(test)]
+mod tests;
