+++ src/values/encoding.rs
@@ -100,97 +100,6 @@ fn to_binary(c: u8) -> u8 {
     }
 }
 
-#[cfg(test)]
-mod binary {
-    use super::{decode_base64, encode_base64};
-    use std::fmt;
-
-    // https://tools.ietf.org/html/rfc4648#section-10
-    #[test]
-    fn encode_rfc4648_test_sample() -> fmt::Result {
-        let mut output = String::with_capacity(8);
-        encode_base64(&mut output, b"")?;
-        assert_eq!(output, "");
-        output.clear();
-
-        encode_base64(&mut output, b"f")?;
-        assert_eq!(output, "Zg==");
-        output.clear();
-
-        encode_base64(&mut output, b"fo")?;
-        assert_eq!(output, "Zm8=");
-        output.clear();
-
-        encode_base64(&mut output, b"foo")?;
-        assert_eq!(output, "Zm9v");
-        output.clear();
-
-        encode_base64(&mut output, b"foob")?;
-        assert_eq!(output, "Zm9vYg==");
-        output.clear();
-
-        encode_base64(&mut output, b"fooba")?;
-        assert_eq!(output, "Zm9vYmE=");
-        output.clear();
-
-        encode_base64(&mut output, b"foobar")?;
-        assert_eq!(output, "Zm9vYmFy");
-        Ok(())
-    }
-
-    #[test]
-    fn encode_text() -> fmt::Result {
-        let input = "Polyfon zwitschernd aßen Mäxchens Vögel Rüben, Joghurt und Quark".as_bytes();
-        let mut output = String::with_capacity(input.len() + input.len() / 3);
-
-        encode_base64(&mut output, input)?;
-        assert_eq!(output, "UG9seWZvbiB6d2l0c2NoZXJuZCBhw59lbiBNw6R4Y2hlbnMgVsO2Z2VsIFLDvGJlbiwgSm9naHVydCB1bmQgUXVhcms=");
-        Ok(())
-    }
-
-    #[test]
-    fn decode_rfc4648_test_sample() {
-        let mut output = Vec::with_capacity(8);
-        decode_base64(&mut output, "");
-        assert_eq!(output, b"");
-        output.clear();
-
-        decode_base64(&mut output, "Zg==");
-        assert_eq!(output, b"f");
-        output.clear();
-
-        decode_base64(&mut output, "Zm8=");
-        assert_eq!(output, b"fo");
-        output.clear();
-
-        decode_base64(&mut output, "Zm9v");
-        assert_eq!(output, b"foo");
-        output.clear();
-
-        decode_base64(&mut output, "Zm9vYg==");
-        assert_eq!(output, b"foob");
-        output.clear();
-
-        decode_base64(&mut output, "Zm9vYmE=");
-        assert_eq!(output, b"fooba");
-        output.clear();
-
-        decode_base64(&mut output, "Zm9vYmFy");
-        assert_eq!(output, b"foobar");
-    }
-
-    #[test]
-    fn decode_text() {
-        let input = "UG9seWZvbiB6d2l0c2NoZXJuZCBhw59lbiBNw6R4Y2hlbnMgVsO2Z2VsIFLDvGJlbiwgSm9naHVydCB1bmQgUXVhcms=";
-        let mut output = Vec::with_capacity(input.len() - input.len() / 3);
-        decode_base64(&mut output, input);
-        assert_eq!(
-            output,
-            "Polyfon zwitschernd aßen Mäxchens Vögel Rüben, Joghurt und Quark".as_bytes()
-        );
-    }
-}
-
 /// Escapes comma, semicolon and backlash character by prepending a backlash.
 ///
 /// This method is used for properties with the value type "TEXT".
@@ -221,34 +130,4 @@ pub(crate) fn escape_text<'t>(text: Cow<'t, str>) -> Cow<'t, str> {
 
 fn is_escaped_char(c: char) -> bool {
     c == ',' || c == ';' || c == '\\'
-}
-
-#[cfg(test)]
-mod text {
-    use super::escape_text;
-
-    #[test]
-    fn escaped_chars() {
-        let s = ",\r\n;:\\ \n \r\n";
-        let expected = "\\,\n\\;:\\\\ \n \n";
-        assert_eq!(expected, escape_text(s.into()));
-    }
-
-    #[test]
-    fn no_escaped_chars() {
-        let s = "This is a simple sentence.";
-        let expected = s.clone();
-        assert_eq!(expected, escape_text(s.into()));
-    }
-
-    // test run with default features enabled but should be correct regardless
-    #[test]
-    fn escape_property() {
-        use components::Property;
-
-        let s = "Hello, World! Today is a beautiful day to test: Escape Methods.\n Characters like ; or \\ must be escaped.\r\n";
-        let expected_value = "Hello\\, World! Today is a beautiful day to test: Escape Methods.\n Characters like \\; or \\\\ must be escaped.\n";
-        let property = Property::new("COMMENT", escape_text(s.into()));
-        assert_eq!(expected_value, property.value);
-    }
 }
