+++ src/values/string.rs
@@ -222,32 +222,3 @@ impl<'t> fmt::Display for Text<'t> {
 //         unimplemented!()
 //     }
 // }
-
-#[cfg(test)]
-mod test {
-    use super::Binary;
-
-    // https://tools.ietf.org/html/rfc4648#section-10
-    #[test]
-    fn parse_valid_binary() {
-        assert_eq!(Some(Binary::new(b"".as_ref())), "".parse().ok());
-        assert_eq!(Some(Binary::new(b"f".as_ref())), "Zg==".parse().ok());
-        assert_eq!(Some(Binary::new(b"fo".as_ref())), "Zm8=".parse().ok());
-        assert_eq!(Some(Binary::new(b"foo".as_ref())), "Zm9v".parse().ok());
-        assert_eq!(Some(Binary::new(b"foob".as_ref())), "Zm9vYg==".parse().ok());
-        assert_eq!(
-            Some(Binary::new(b"fooba".as_ref())),
-            "Zm9vYmE=".parse().ok()
-        );
-        assert_eq!(
-            Some(Binary::new(b"foobar".as_ref())),
-            "Zm9vYmFy".parse().ok()
-        );
-    }
-
-    #[test]
-    fn parse_invalid_binary() {
-        assert!("ABC".parse::<Binary>().is_err());
-        assert!("Zö==".parse::<Binary>().is_err());
-    }
-}
