+++ src/values/time.rs
@@ -598,40 +598,3 @@ impl fmt::Display for Time<Utc> {
 // }
 
 // pub struct Recur;
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn time_utc() {
-        let expected = "173000Z";
-        let time = Time::utc(17, 30, 0).unwrap();
-        assert_eq!(time.to_string(), expected);
-    }
-
-    #[test]
-    fn time_local() {
-        let expected = "173000";
-        let time = Time::local(17, 30, 0).unwrap();
-        assert_eq!(time.to_string(), expected);
-    }
-
-    #[test]
-    fn datetime_local() {
-        let expected = "19970714T173000";
-        let date = Date::new(1997, Month::July, 14).unwrap();
-        let time = Time::local(17, 30, 0).unwrap();
-        let datetime = DateTime::new(date, time);
-        assert_eq!(datetime.to_string(), expected);
-    }
-
-    #[test]
-    fn datetime_utc() {
-        let expected = "19970714T173000Z";
-        let date = Date::new(1997, Month::July, 14).unwrap();
-        let time = Time::utc(17, 30, 0).unwrap();
-        let datetime = DateTime::new(date, time);
-        assert_eq!(datetime.to_string(), expected);
-    }
-}
